add function to add/edit users in department

+++ src/app/components/requests/admin/admin-req-dept/admin-req-dept.component.ts
@@ -32,11 +32,15 @@ export class AdminReqDeptComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe(dept => {
       if(dept.deptmt_name != null && dept.deptmt_name.trim().length > 0){
+        document.querySelector("body").style.cssText = "cursor: wait";
         this.requestService.addRequesterDept(dept).subscribe(result => {
+          document.querySelector("body").style.cssText = "cursor: auto";
           if(result.length >= 0){
             this.authService.departments = result;
             this.reqDeptDataSource = new MatTableDataSource(result);
             this.reqDeptDataSource.sort = this.sort;
+
+            this.updateUser();
           }
           else if(result.ok == false){
             const dialogRef = this.dialog.open(NotificationDialog, { data: "Error: " + result.message, width: '600px'});
@@ -51,11 +55,14 @@ export class AdminReqDeptComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe(dept => {
       if(dept.deptmt_name != null && dept.deptmt_name.trim().length > 0){
+        document.querySelector("body").style.cssText = "cursor: wait";
         this.requestService.editRequesterDept(dept).subscribe(result => {
+          document.querySelector("body").style.cssText = "cursor: auto";
           if(result.length >= 0){
             this.authService.departments = result;
             this.reqDeptDataSource = new MatTableDataSource(result);
             this.reqDeptDataSource.sort = this.sort;
+            this.updateUser();
           }
           else if(result.ok == false){
             const dialogRef = this.dialog.open(NotificationDialog, { data: "Error: " + result.message, width: '600px'});
@@ -66,21 +73,35 @@ export class AdminReqDeptComponent implements OnInit {
   }
 
   deleteReqDept(element:any){
-    const dialogRef = this.dialog.open(ConfirmationDialog, { data: {title: "Delete Requester Dept Confirmation", message: "Are you sure you want to delete this requester department?"}, width: '600px'});
+    if(element.users != null && element.users.length > 0){
+      const dialogRef = this.dialog.open(NotificationDialog, { data: "There are users associated with the department. Please remove the users before proceeding", width: '600px'});
+    }
+    else{
+      const dialogRef = this.dialog.open(ConfirmationDialog, { data: {title: "Delete Requester Dept Confirmation", message: "Are you sure you want to delete this requester department?"}, width: '600px'});
 
-    dialogRef.afterClosed().subscribe(result => {
-      if(result){
-        this.requestService.deleteRequesterDept(element.deptmt_id).subscribe(result => {
-          if(result.length >= 0){
-            this.authService.departments = result;
-            this.reqDeptDataSource = new MatTableDataSource(result);
-            this.reqDeptDataSource.sort = this.sort;
-          }
-          else if(result.ok == false){
-            const dialogRef = this.dialog.open(NotificationDialog, { data: "Error: " + result.message, width: '600px'});
-          }
-        });
-      }
+      dialogRef.afterClosed().subscribe(result => {
+        if(result){
+          document.querySelector("body").style.cssText = "cursor: wait";
+          this.requestService.deleteRequesterDept(element.deptmt_id).subscribe(result => {
+            document.querySelector("body").style.cssText = "cursor: auto";
+            if(result.length >= 0){
+              this.authService.departments = result;
+              this.reqDeptDataSource = new MatTableDataSource(result);
+              this.reqDeptDataSource.sort = this.sort;
+              this.updateUser();
+            }
+            else if(result.ok == false){
+              const dialogRef = this.dialog.open(NotificationDialog, { data: "Error: " + result.message, width: '600px'});
+            }
+          });
+        }
+      });
+    }
+  }
+
+  updateUser(){
+    this.domainService.getAllUsers(this.authService.appSettings.service_url).subscribe(result => {
+      this.authService.allUsers = result;
     });
   }
 }
